+++ src/models/index.ts
@@ -8,6 +8,7 @@ import type SerializesToJsonLD from './mixins/SerializesToJsonLD';
 import SolidContainerModel from './SolidContainerModel';
 import SolidDocument from './SolidDocument';
 import SolidModelMetadata from './SolidModelMetadata';
+import SolidModelOperation from './SolidModelOperation';
 
 export * from './inference';
 export * from './relations/index';
@@ -17,6 +18,7 @@ export {
     SolidContainerModel,
     SolidDocument,
     SolidModelMetadata,
+    SolidModelOperation,
 };
 
 export type {
@@ -32,5 +34,5 @@ export type {
 } from './fields';
 
 export function bootSolidModels(): void {
-    bootModels({ SolidDocument, SolidModelMetadata });
+    bootModels({ SolidDocument, SolidModelMetadata, SolidModelOperation });
 }